requirements-dev.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile requirements-dev.in
#
asgiref==3.6.0
    # via
    #   -r requirements.txt
    #   django
attrs==22.2.0
    # via
    #   flake8-bugbear
    #   pytest
build==0.10.0
    # via pip-tools
certifi==2022.12.7
    # via requests
charset-normalizer==3.0.1
    # via requests
classify-imports==4.2.0
    # via flake8-type-checking
click==8.1.3
    # via pip-tools
coverage[toml]==7.1.0
    # via pytest-cov
crispy-bulma==0.8.3
    # via -r requirements.txt
django==3.2.17
    # via
    #   -r requirements.txt
    #   crispy-bulma
    #   django-filter
    #   django-stubs
    #   django-stubs-ext
django-crispy-forms==1.14.0
    # via
    #   -r requirements.txt
    #   crispy-bulma
django-filter==22.1
    # via -r requirements.txt
django-filter-stubs==0.1.2
    # via -r requirements-dev.in
django-stubs[compatible-mypy]==1.14.0
    # via
    #   -r requirements-dev.in
    #   django-filter-stubs
    #   djangorestframework-stubs
django-stubs-ext==0.7.0
    # via
    #   -r requirements.txt
    #   django-stubs
djangorestframework-stubs==1.8.0
    # via django-filter-stubs
exceptiongroup==1.1.0
    # via pytest
faker==16.7.0
    # via -r requirements-dev.in
flake8==6.0.0
    # via
    #   -r requirements-dev.in
    #   flake8-bugbear
    #   flake8-builtins
    #   flake8-comprehensions
    #   flake8-debugger
    #   flake8-isort
    #   flake8-mutable
    #   flake8-type-checking
flake8-bugbear==23.1.20
    # via -r requirements-dev.in
flake8-builtins==2.1.0
    # via -r requirements-dev.in
flake8-comprehensions==3.10.1
    # via -r requirements-dev.in
flake8-debugger==4.1.2
    # via -r requirements-dev.in
flake8-isort==6.0.0
    # via -r requirements-dev.in
flake8-mutable==1.2.0
    # via -r requirements-dev.in
flake8-todo==0.7
    # via -r requirements-dev.in
flake8-type-checking==2.3.0
    # via -r requirements-dev.in
freezegun==1.2.2
    # via -r requirements-dev.in
idna==3.4
    # via requests
iniconfig==2.0.0
    # via pytest
isort==5.12.0
    # via
    #   -r requirements-dev.in
    #   flake8-isort
mccabe==0.7.0
    # via flake8
mypy==0.991
    # via
    #   -r requirements-dev.in
    #   django-filter-stubs
    #   django-stubs
    #   djangorestframework-stubs
mypy-extensions==0.4.3
    # via mypy
packaging==23.0
    # via
    #   build
    #   pytest
pip-tools==6.12.2
    # via -r requirements-dev.in
pluggy==1.0.0
    # via pytest
pycodestyle==2.10.0
    # via
    #   flake8
    #   flake8-debugger
    #   flake8-todo
pyflakes==3.0.1
    # via flake8
pyproject-hooks==1.0.0
    # via build
pytest==7.2.1
    # via
    #   -r requirements-dev.in
    #   pytest-cov
    #   pytest-django
pytest-cov==4.0.0
    # via -r requirements-dev.in
pytest-django==4.5.2
    # via -r requirements-dev.in
python-dateutil==2.8.2
    # via
    #   faker
    #   freezegun
pytz==2022.7.1
    # via
    #   -r requirements.txt
    #   django
requests==2.28.2
    # via djangorestframework-stubs
six==1.16.0
    # via python-dateutil
sqlparse==0.4.3
    # via
    #   -r requirements.txt
    #   django
tomli==2.0.1
    # via
    #   build
    #   coverage
    #   django-stubs
    #   mypy
    #   pyproject-hooks
    #   pytest
types-docutils==0.19.1.2
    # via types-setuptools
types-pytz==2022.7.1.0
    # via django-stubs
types-pyyaml==6.0.12.3
    # via
    #   django-stubs
    #   djangorestframework-stubs
types-requests==2.28.11.12
    # via djangorestframework-stubs
types-setuptools==65.7.0.3
    # via -r requirements-dev.in
types-urllib3==1.26.25.5
    # via types-requests
typing-extensions==4.4.0
    # via
    #   -r requirements.txt
    #   django-filter-stubs
    #   django-stubs
    #   django-stubs-ext
    #   djangorestframework-stubs
    #   mypy
urllib3==1.26.14
    # via requests
wheel==0.38.4
    # via pip-tools

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
